+++ CHANGELOG.md
@@ -27,7 +27,7 @@ This project adheres to the Node [default version scheme](https://docs.npmjs.com
 - Internal: Introduce integration tests for API endpoint integrations.
 - Internal: Added ScreenLayout component. This is currently used in the Welcome and Complete screens.
 - Internal: Introduce TypeScript on non-critical components & deprecate FlowType.
-- Internal: App component, Redux system, utils & HoCs are now typed in TypeScript.
+- Internal: App component, Redux system, utils, HoCs & routers are now typed in TypeScript.
 
 ### Changed